Split the core functions of the course browsing page into different components, including: course cards, filters, and paging components
All relevant interface definitions are moved to the src/types/ directory and are reasonably classified and organized by function.
