+++ src/components/browse/CourseFilters.tsx
@@ -1,0 +1,100 @@
+'use client';
+
+import React from 'react';
+import { ChevronUpDownIcon } from '@/components/ui/Icons';
+import { CourseFiltersProps } from '@/types/components';
+
+export default function CourseFilters({
+  categories,
+  selectedCategory,
+  priceRange,
+  sortBy,
+  onCategoryChange,
+  onPriceRangeChange,
+  onSortChange,
+  onClearFilters,
+}: CourseFiltersProps) {
+  const allCategories = categories.flatMap((cat) => cat.category);
+  const uniqueCategories = Array.from(new Set(allCategories));
+
+  return (
+    <div className="mt-6 grid grid-cols-1 md:grid-cols-4 gap-4">
+      {/* Category Filter */}
+      <div>
+        <label className="block text-sm font-medium text-gray-700 mb-2">
+          Category
+        </label>
+        <div className="relative">
+          <select
+            value={selectedCategory}
+            onChange={(e) => onCategoryChange(e.target.value)}
+            className="block w-full pl-3 pr-10 py-2 text-base border-gray-300 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm rounded-md"
+          >
+            <option value="all">All Categories</option>
+            {uniqueCategories.map((category) => (
+              <option key={category} value={category}>
+                {category}
+              </option>
+            ))}
+          </select>
+          <ChevronUpDownIcon className="absolute right-3 top-2.5 h-5 w-5 text-gray-400 pointer-events-none" />
+        </div>
+      </div>
+
+      {/* Price Range Filter */}
+      <div>
+        <label className="block text-sm font-medium text-gray-700 mb-2">
+          Price Range ($)
+        </label>
+        <div className="flex space-x-2">
+          <input
+            type="number"
+            placeholder="Min"
+            value={priceRange.min}
+            onChange={(e) => onPriceRangeChange('min', e.target.value)}
+            className="block w-full px-3 py-2 text-base border-gray-300 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm rounded-md"
+          />
+          <input
+            type="number"
+            placeholder="Max"
+            value={priceRange.max}
+            onChange={(e) => onPriceRangeChange('max', e.target.value)}
+            className="block w-full px-3 py-2 text-base border-gray-300 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm rounded-md"
+          />
+        </div>
+      </div>
+
+      {/* Sort By */}
+      <div>
+        <label className="block text-sm font-medium text-gray-700 mb-2">
+          Sort By
+        </label>
+        <div className="relative">
+          <select
+            value={sortBy}
+            onChange={(e) => onSortChange(e.target.value)}
+            className="block w-full pl-3 pr-10 py-2 text-base border-gray-300 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm rounded-md"
+          >
+            <option value="newest">Newest First</option>
+            <option value="oldest">Oldest First</option>
+            <option value="price-low">Price: Low to High</option>
+            <option value="price-high">Price: High to Low</option>
+            <option value="rating">Highest Rated</option>
+            <option value="title">Title A-Z</option>
+          </select>
+          <ChevronUpDownIcon className="absolute right-3 top-2.5 h-5 w-5 text-gray-400 pointer-events-none" />
+        </div>
+      </div>
+
+      {/* Clear Filters */}
+      <div className="flex items-end">
+        <button
+          onClick={onClearFilters}
+          className="w-full px-4 py-2 text-sm font-medium text-gray-700 bg-white border border-gray-300 rounded-md hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
+        >
+          Clear Filters
+        </button>
+      </div>
+    </div>
+  );
+}
