+++ src/components/browse/CoursePagination.tsx
@@ -1,0 +1,38 @@
+'use client';
+
+import React from 'react';
+import { CoursePaginationProps } from '@/types/components';
+
+export default function CoursePagination({
+  currentPage,
+  totalPages,
+  onPageChange,
+}: CoursePaginationProps) {
+  if (totalPages <= 1) return null;
+
+  return (
+    <div className="mt-8 flex justify-center">
+      <nav className="flex items-center space-x-2">
+        <button
+          onClick={() => onPageChange(Math.max(currentPage - 1, 1))}
+          disabled={currentPage === 1}
+          className="px-3 py-2 text-sm font-medium text-gray-500 bg-white border border-gray-300 rounded-md hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed"
+        >
+          Previous
+        </button>
+
+        <span className="px-3 py-2 text-sm font-medium text-gray-700">
+          Page {currentPage} of {totalPages}
+        </span>
+
+        <button
+          onClick={() => onPageChange(Math.min(currentPage + 1, totalPages))}
+          disabled={currentPage === totalPages}
+          className="px-3 py-2 text-sm font-medium text-gray-500 bg-white border border-gray-300 rounded-md hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed"
+        >
+          Next
+        </button>
+      </nav>
+    </div>
+  );
+}
