+++ src/types/course.ts
@@ -34,6 +34,61 @@ export interface Course {
   createdAt: string;
 }
 
+// Course with detailed modules for course detail page
+export interface CourseWithModules {
+  _id: string;
+  title: string;
+  description: string;
+  thumbnail?: string;
+  modules: Module[];
+  instructor: {
+    name: string;
+    email: string;
+  };
+  category: string;
+  level: string;
+  averageRating: number;
+}
+
+// Course with populated instructor for browse page
+export interface CourseWithInstructor {
+  _id: string;
+  title: string;
+  description: string;
+  thumbnail?: string;
+  modules: Array<{
+    _id: string;
+    title: string;
+    duration: number;
+    order: number;
+  }>;
+  instructor: {
+    _id: string;
+    name: string;
+    email: string;
+  };
+  category: string;
+  level: string;
+  averageRating: number;
+  price: number;
+  createdAt: string;
+}
+
+// Module interface
+export interface Module {
+  _id: string;
+  title: string;
+  content: string;
+  duration: number;
+  order: number;
+}
+
+// Category interface
+export interface Category {
+  _id: string;
+  category: string[];
+}
+
 // API related types for courses route
 export interface CourseQuery {
   status: string;
